requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
openpyxl==3.1.2
werkzeug==3.0.1
python-dateutil==2.8.2